Выполняет домашнее задание 5 по модуляции
