+++ js/util.js
@@ -1,0 +1,14 @@
+const getRandomIntInclusive = function (a, b) {
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+
+  const min = Math.ceil(Math.min(a, b));
+  const max = Math.floor(Math.max(a, b));
+
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+export {
+  getRandomIntInclusive
+};
